Added did delete files
